big commit with a bunch of improvements from the private code branch. restructured units. added lots of sink particle and single-star functionality, added lots of arbitrary-EOS functionality, added new cooling and TREECOOL tables, updated dark matter modules, updated default sets for parameter files, removed a huge number of extraneous parameters and config flags, continued restructuring of routines to be more fully-modular. standard bug-fixes and performance improvements.

+++ structure/subfind/subfind_density.c
@@ -277,8 +277,8 @@ void subfind_density(int j_in)
 		  if(iter >= MAXITER - 10)
 		    {
 		      printf
-			("i=%d task=%d ID=%d Hsml=%g Left=%g Right=%g Ngbs=%g Right-Left=%g\n   pos=(%g|%g|%g)\n",
-			 i, ThisTask, (int) P[i].ID, P[i].DM_Hsml, Left[i], Right[i],
+			("i=%d task=%d ID=%llu Hsml=%g Left=%g Right=%g Ngbs=%g Right-Left=%g\n   pos=(%g|%g|%g)\n",
+			 i, ThisTask, (unsigned long long) P[i].ID, P[i].DM_Hsml, Left[i], Right[i],
 			 (double) P[i].DM_NumNgb, Right[i] - Left[i], P[i].Pos[0], P[i].Pos[1], P[i].Pos[2]);
 		      fflush(stdout);
 		    }
